+++ web/src/features/terminal-ticket/form-header-filter.tsx
@@ -3,7 +3,7 @@ import { SearchIcon } from 'lucide-react';
 import Box from '@/components/box';
 import { Fieldset } from '@/components/fieldset';
 import { Flex, FlexCol } from '@/components/flex';
-import { Button } from '@/components/ui/button';
+import { IconButton } from '@/components/button/IconButton';
 import { Input } from '@/components/ui/input.tsx';
 import {
   Select,
@@ -13,7 +13,7 @@ import {
   SelectValue,
 } from '@/components/ui/select.tsx';
 import { TypographyMuted } from '@/components/ui/typography-muted.tsx';
-import { SelectDayToSearch } from '@/features/plays-and-hits/select-day-to-search.tsx';
+import { SelectDayToSearch } from '@/components/button/SelectDayToSearch';
 
 import IsRoleCashier from '@/components/is-role-cashier';
 import { useEffect, useState } from 'react';
@@ -21,6 +21,7 @@ import { useUsers } from '@/hooks/fetchs/users/useUsers';
 import dayjs from 'dayjs';
 import { useSearchParams } from 'react-router-dom';
 import { useAuth } from '@/contexts/AuthContext';
+import { USER_TYPE } from '@helper/types/user.type';
 
 const FormHeaderFilter = () => {
   const { role } = useAuth();
@@ -63,96 +64,106 @@ const FormHeaderFilter = () => {
     handleSelectDate();
   }, []);
   return (
-    <Flex className={' space-y-4 sm:w-full'}>
-      <FlexCol className={'mb-1 sm:mb-4 1440:gap-6  gap-2 '}>
-        <Fieldset legend={'Pasador:'} className={'sm:w-full gap-1 sm:gap-3'}>
-          <FlexCol className={'space-y-4'}>
-            <IsRoleCashier role={role}>
-              <Flex className={' gap-3'}>
+    <FlexCol className={'sm:flex-row mb-1 sm:mb-4 w-full gap-1 sm:gap-3'}>
+      <Fieldset
+        legend={role === USER_TYPE.CASHIER ? 'Seleccionar fecha' : 'Pasador:'}
+        className={'w-full sm:w-1/2 flex gap-1 sm:gap-3'}
+      >
+        <FlexCol className={'w-full  space-y-4'}>
+          <IsRoleCashier role={role}>
+            <Flex className={' gap-3'}>
+              <Select
+                defaultValue={undefined}
+                value={selectValue}
+                onValueChange={(value) => {
+                  handleSelectCashier(value);
+                }}
+              >
+                <SelectTrigger className={'border w-full '}>
+                  <SelectValue placeholder="Todos" />
+                </SelectTrigger>
+                <SelectContent>
+                  {cashiers?.map((cashier) => {
+                    return (
+                      <SelectItem key={cashier.user_id} value={cashier.user_id}>
+                        {cashier.name} - {cashier.number}
+                      </SelectItem>
+                    );
+                  })}
+                </SelectContent>
+              </Select>
+            </Flex>
+          </IsRoleCashier>
+
+          <Flex className={'gap-2 w-full sm:w-2/4 justify-between'}>
+            <Flex className={'items-center gap-3'}>
+              <Box className="hidden sm:box">
+                <TypographyMuted label={'Fecha'} />
+              </Box>
+              <Box className={' overflow-hidden'}>
+                <SelectDayToSearch
+                  onDayChange={(value) => {
+                    handleSelectDate(value);
+                  }}
+                  toDate={dayjs().toDate()}
+                />
+              </Box>
+            </Flex>
+            <Flex>
+              <Flex className={'w-full items-center gap-3'}>
+                <Box>
+                  <TypographyMuted label={'Tickets'} />
+                </Box>
                 <Select
-                  defaultValue={undefined}
-                  value={selectValue}
-                  onValueChange={(value) => {
-                    handleSelectCashier(value);
+                  onValueChange={(value: 'all' | 'winner' | 'paid' | 'not_paid') => {
+                    onChangeFilter(value);
                   }}
                 >
-                  <SelectTrigger className={'border w-full '}>
+                  <SelectTrigger className="border w-full">
                     <SelectValue placeholder="Todos" />
                   </SelectTrigger>
                   <SelectContent>
-                    {cashiers?.map((cashier) => {
-                      return (
-                        <SelectItem key={cashier.user_id} value={cashier.user_id}>
-                          {cashier.name} - {cashier.number}
-                        </SelectItem>
-                      );
-                    })}
+                    <SelectItem value="all">Todos</SelectItem>
+                    <SelectItem value="winner">Con aciertos</SelectItem>
+                    <SelectItem value="paid">Pagados</SelectItem>
+                    <SelectItem value="not_paid">No Pagados</SelectItem>
                   </SelectContent>
                 </Select>
               </Flex>
-            </IsRoleCashier>
-
-            <Flex className={'flex-1 gap-2 '}>
-              <Flex className={'items-center gap-3'}>
-                <Box className="hidden sm:box">
-                  <TypographyMuted label={'Fecha'} />
-                </Box>
-                <Box className={'sm:w-[200px] overflow-hidden'}>
-                  <SelectDayToSearch
-                    onDayChange={(value) => {
-                      handleSelectDate(value);
-                    }}
-                    toDate={dayjs().toDate()}
-                  />
-                </Box>
-              </Flex>
-              <Flex className={'w-[150px]'}>
-                <Flex className={'w-full items-center gap-3'}>
-                  <Box>
-                    <TypographyMuted label={'Tickets'} />
-                  </Box>
-                  <Select
-                    onValueChange={(value: 'all' | 'winner' | 'paid' | 'not_paid') => {
-                      onChangeFilter(value);
-                    }}
-                  >
-                    <SelectTrigger className="border w-full">
-                      <SelectValue placeholder="Todos" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="all">Todos</SelectItem>
-                      <SelectItem value="winner">Con aciertos</SelectItem>
-                      <SelectItem value="paid">Pagados</SelectItem>
-                      <SelectItem value="not_paid">No Pagados</SelectItem>
-                    </SelectContent>
-                  </Select>
-                </Flex>
-              </Flex>
-            </Flex>
-          </FlexCol>
-        </Fieldset>
-
-        <Fieldset legend="Buscar por número de Ticket:" className="w-full gap-3">
-          <Flex className="gap-4">
-            <Input
-              type="number"
-              inputMode="numeric"
-              value={inputValue}
-              onChange={(e) => setInputValue(e.target.value)}
-              placeholder="Ej: 20250619..."
-            />
-            <Flex className="gap-4">
-              <Button type="button" className="!px-6" onClick={handleSearch}>
-                <SearchIcon /> Buscar
-              </Button>
-              <Button type="reset" variant="outline" className="!px-6" onClick={handleReset}>
-                Limpiar
-              </Button>
             </Flex>
           </Flex>
-        </Fieldset>
-      </FlexCol>
-    </Flex>
+        </FlexCol>
+      </Fieldset>
+
+      <Fieldset legend="Buscar por número de Ticket:" className="w-full sm:w-1/2 flex gap-3">
+        <FlexCol className="w-full gap-4">
+          <Input
+            type="number"
+            inputMode="numeric"
+            className=""
+            value={inputValue}
+            onChange={(e) => setInputValue(e.target.value)}
+            placeholder="Ej: 20250619..."
+          />
+          <Flex className="gap-4 w-full justify-between">
+            <IconButton
+              type="button"
+              label="Buscar"
+              icon={<SearchIcon />}
+              onClick={handleSearch}
+              className="!px-6"
+            />
+            <IconButton
+              type="reset"
+              label="Limpiar"
+              variant="outline"
+              onClick={handleReset}
+              className="!px-6"
+            />
+          </Flex>
+        </FlexCol>
+      </Fieldset>
+    </FlexCol>
   );
 };