+++ web/src/features/upcoming-lotteries/schedules-list.tsx
@@ -1,8 +1,6 @@
-import { Label } from '@/components/ui/label';
 import SkeletonList from '@/components/skeletons/skeleton-list';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
-
-import { Flex } from '@/components/flex';
+import { RadioGroup } from '@/components/ui/radio-group';
+import { RadioButtonWithLabel } from '@/components/button/RadioButtonWithLabel';
 import { IScheduleEntityFront } from '@helper/types/schedule.type';
 import { useSchedules } from '@/hooks/fetchs/schedule/useSchedules';
 
@@ -29,19 +27,13 @@ export const ScheduleRadioList = ({ selectedSchedule, handleSchedule }: Schedule
       value={selectedSchedule}
     >
       {schedules?.map((schedule: IScheduleEntityFront) => (
-        <Flex className={'gap-3 items-center'} key={schedule.schedule_id}>
-          <RadioGroupItem
-            id={schedule.schedule_id}
-            value={schedule.schedule_id}
-            className="border-2 border-primary"
-          />
-          <Label
-            htmlFor={schedule.schedule_id}
-            className="1440:text-sm text-xs font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-          >
-            {schedule.name} ({schedule.time.slice(0, 5)})
-          </Label>
-        </Flex>
+        <RadioButtonWithLabel
+          key={schedule.schedule_id}
+          id={schedule.schedule_id}
+          value={schedule.schedule_id}
+          label={`${schedule.name} (${schedule.time.slice(0, 5)})`}
+          labelClassName="1440:text-sm text-xs"
+        />
       ))}
     </RadioGroup>
   );